requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
pytest
pydantic-settings
python-multipart
pandas
pathlib
openai
openai==0.28
black
isort
pylint
boto3
pytest
pytest-mock
